fix: moved ridge tests from test_linear to test_ridge
